Conflicts:
	src/webapp/components/uiOptions/js/FatPanelUIOptions.js
	src/webapp/components/uiOptions/js/SettingsPanels.js
	src/webapp/tests/component-tests/uiOptions/js/FatPanelUIOptionsTests.js

+++ src/webapp/framework/core/js/FluidIoC.js
@@ -145,7 +145,7 @@ var fluid_1_5 = fluid_1_5 || {};
     };
 
     // unsupported, NON-API function
-    fluid.makeDistributionRecord = function (sourceRecord, sourcePath, targetPath, exclusions, offset, sourceType) {
+    fluid.makeDistributionRecord = function (contextThat, sourceRecord, sourcePath, targetSegs, exclusions, offset, sourceType) {
         offset = offset || 0;
         sourceType = sourceType || "distribution";
 
@@ -154,20 +154,22 @@ var fluid_1_5 = fluid_1_5 || {};
             fluid.model.applyChangeRequest(source, {path: exclusion, type: "DELETE"});
         });
 
-        var options = {};        
-        fluid.model.applyChangeRequest(options, {path: targetPath, type: "ADD", value: source});
-        return {options: options, recordType: sourceType, priority: fluid.mergeRecordTypes.distribution + offset};
+        var record = {options: {}};
+        fluid.model.applyChangeRequest(record, {path: targetSegs, type: "MERGE", value: source});
+        return $.extend(record, {contextThat: contextThat, recordType: sourceType, priority: fluid.mergeRecordTypes.distribution + offset});
     };
 
     // unsupported, NON-API function    
-    fluid.filterBlocks = function (sourceBlocks, sourcePath, targetPath, exclusions) {
+    fluid.filterBlocks = function (contextThat, sourceBlocks, sourcePath, targetSegs, exclusions, removeSource) {
         var togo = [], offset = 0;
         fluid.each(sourceBlocks, function (block) {
             var source = fluid.get(block.source, sourcePath);
             if (source) {
-                togo.push(fluid.makeDistributionRecord(block.source, sourcePath, targetPath, exclusions, offset++, block.recordType));
+                togo.push(fluid.makeDistributionRecord(contextThat, block.source, sourcePath, targetSegs, exclusions, offset++, block.recordType));
                 var rescued = $.extend({}, source);
-                fluid.model.applyChangeRequest(block.source, {path: sourcePath, type: "DELETE"});
+                if (removeSource) {
+                    fluid.model.applyChangeRequest(block.source, {path: sourcePath, type: "DELETE"});
+                }
                 fluid.each(exclusions, function (exclusion) {
                     var orig = fluid.get(rescued, exclusion);
                     fluid.set(block.source, sourcePath.concat(exclusion), orig);
@@ -178,7 +180,7 @@ var fluid_1_5 = fluid_1_5 || {};
     };
     
     // unsupported, NON-API function    
-    fluid.matchIoCSelector = function (selector, thatStack, shadows, memberNames, i) {
+    fluid.matchIoCSelector = function (selector, thatStack, contextHashes, memberNames, i) {
         var thatpos = thatStack.length - 1;
         var selpos = selector.length - 1;
         while (true) {
@@ -189,7 +191,7 @@ var fluid_1_5 = fluid_1_5 || {};
             var match = true;
             for (var j = 0; j < selel.predList.length; ++j) {
                 var pred = selel.predList[j];
-                if (pred.context && !(shadows[thatpos].contextHash[pred.context] || memberNames[thatpos] === pred.context)) {
+                if (pred.context && !(contextHashes[thatpos][pred.context] || memberNames[thatpos] === pred.context)) {
                     match = false;
                     break;
                 }
@@ -220,27 +222,37 @@ var fluid_1_5 = fluid_1_5 || {};
     };
     
     // unsupported, NON-API function
-    fluid.collectDistributions = function (distributedBlocks, distribution, thatStack, shadows, memberNames, i) {
-        if (fluid.matchIoCSelector(distribution.selector, thatStack, shadows, memberNames, i)) {
+    fluid.collectDistributions = function (distributedBlocks, distribution, thatStack, contextHashes, memberNames, i) {
+        if (fluid.matchIoCSelector(distribution.selector, thatStack, contextHashes, memberNames, i)) {
             distributedBlocks.push.apply(distributedBlocks, distribution.blocks);
         }
     };
 
+    fluid.getMembers = function (holder, name) {
+        return fluid.transform(holder, function(member) {
+            return fluid.get(member, name);
+        });
+    };
+
     // unsupported, NON-API function 
-    fluid.receiveDistributions = function (that) {
-        var instantiator = fluid.getInstantiator(that);
-        var thatStack = instantiator.getThatStack(that); // most specific is at end
+    fluid.receiveDistributions = function (parentThat, gradeNames, memberName) {
+        var instantiator = fluid.getInstantiator(parentThat);
+        var thatStack = instantiator.getThatStack(parentThat); // most specific is at end
         var memberNames = fluid.getMemberNames(instantiator, thatStack);
+        memberNames.push(memberName);
         var distributedBlocks = [];
         var shadows = fluid.transform(thatStack, function (thisThat) {
-            return instantiator.idToShadow[thisThat.id];  
+            return instantiator.idToShadow[thisThat.id];
         });
-        for (var i = 0; i < thatStack.length; ++ i) {
+        var contextHashes = fluid.getMembers(shadows, "contextHash");
+        contextHashes.push(fluid.gradeNamesToHash(gradeNames));
+        thatStack.push({}); // fake "that" since we try to match selectors before creation for FLUID-5013
+        for (var i = 0; i < thatStack.length - 1; ++ i) {
             fluid.each(shadows[i].distributions, function (distribution) {
-                fluid.collectDistributions(distributedBlocks, distribution, thatStack, shadows, memberNames, i);
+                fluid.collectDistributions(distributedBlocks, distribution, thatStack, contextHashes, memberNames, i);
             });
         }
-        fluid.applyDistributions(that, distributedBlocks, shadow);
+        return distributedBlocks;
     };
 
     // unsupported, NON-API function 
@@ -294,7 +306,7 @@ var fluid_1_5 = fluid_1_5 || {};
         var predList = parsedSelector[0].predList;
         var context = predList[0].context;
         predList.length = 0;
-        return context; 
+        return context;
     };
 
     fluid.undistributableOptions = ["gradeNames", "distributeOptions", "returnedPath", "argumentMap", "initFunction", "mergePolicy", "progressiveCheckerOptions"]; // automatically added to "exclusions" of every distribution
@@ -319,10 +331,10 @@ var fluid_1_5 = fluid_1_5 || {};
                     fluid.fail("Error in options distribution record ", record, " - could not resolve context selector {"+targetRef.context+"} to a root component");
                 }
             }
-            var segs = fluid.parseExpectedOptionsPath(targetRef.path, "target");
+            var targetSegs = fluid.model.parseEL(targetRef.path);
             var preBlocks;
             if (record.record) {
-                preBlocks = [(fluid.makeDistributionRecord(record.record, [], segs, [], 0))];
+                preBlocks = [(fluid.makeDistributionRecord(that, record.record, [], targetSegs, [], 0))];
             }
             else {
                 var thatShadow = fluid.shadowForComponent(that);
@@ -331,13 +343,13 @@ var fluid_1_5 = fluid_1_5 || {};
                     fluid.fail("Error in options distribution record ", record, " only a context of {that} is supported");
                 }
                 var sourcePath = fluid.parseExpectedOptionsPath(source.path, "source");
-                var fullExclusions = record.exclusions.concat(sourcePath.length === 0 ? fluid.undistributableOptions : []); 
+                var fullExclusions = fluid.makeArray(record.exclusions).concat(sourcePath.length === 0 ? fluid.undistributableOptions : []); 
                 
                 var exclusions = fluid.transform(fullExclusions, function (exclusion) {
                     return fluid.model.parseEL(exclusion);  
                 });
                 
-                preBlocks = fluid.filterBlocks(thatShadow.mergeOptions.mergeBlocks, sourcePath, segs, exclusions);
+                preBlocks = fluid.filterBlocks(that, thatShadow.mergeOptions.mergeBlocks, sourcePath, targetSegs, exclusions, record.removeSource);
                 thatShadow.mergeOptions.updateBlocks(); // perhaps unnecessary
             }
             // TODO: inline material has to be expanded in its original context!
@@ -353,12 +365,18 @@ var fluid_1_5 = fluid_1_5 || {};
     };
 
     // unsupported, NON-API function    
-    fluid.cacheShadowGrades = function (that, shadow) {
+    fluid.gradeNamesToHash = function (gradeNames) {
         var contextHash = {};
-        fluid.each(that.options.gradeNames, function (gradeName) {
+        fluid.each(gradeNames, function (gradeName) {
             contextHash[gradeName] = true;
             contextHash[fluid.computeNickName(gradeName)] = true;  
         });
+        return contextHash;      
+    };
+
+    // unsupported, NON-API function    
+    fluid.cacheShadowGrades = function (that, shadow) {
+        var contextHash = fluid.gradeNamesToHash(that.options.gradeNames);
         contextHash[that.nickName] = true;
         shadow.contextHash = contextHash;      
     };
@@ -370,9 +388,6 @@ var fluid_1_5 = fluid_1_5 || {};
         var shadow = fluid.shadowForComponent(that, shadow);
         fluid.cacheShadowGrades(that, shadow);
         shadow.mergeOptions = mergeOptions;
-
-        fluid.receiveDistributions(that);
-        fluid.distributeOptions(that, mergeOptions.strategy);
     };
 
     // unsupported, NON-API function    
@@ -381,9 +396,12 @@ var fluid_1_5 = fluid_1_5 || {};
         // This prevents corruption of instantiator records by defeating effect of "returnedPath" for non-roots 
         return shadow && shadow.path !== "" ? null : returnedPath;
     };
-    
+
+    // unsupported, NON-API function    
     fluid.computeDynamicGrades = function (that, shadow, strategy) {
-        var gradeNames = that.options.gradeNames;
+        delete that.options.gradeNames; // Recompute gradeNames for FLUID-5012 and others
+        var gradeNames = fluid.driveStrategy(that.options, "gradeNames", strategy);
+        
         // TODO: In complex distribution cases, a component might end up with multiple default blocks
         var defaultsBlock = fluid.findMergeBlocks(shadow.mergeOptions.mergeBlocks, "defaults")[0];
         var dynamicGrades = fluid.remove_if(gradeNames, function (gradeName) {
@@ -403,7 +421,7 @@ var fluid_1_5 = fluid_1_5 || {};
             fluid.remove_if(otherGrades, function (gradeName) {
                 return gradeName === that.typeName;
             }); // Remove our own typeName which getGradedDefaults is not expecting (FLUID-4939)
-            var newDefaults = fluid.getGradedDefaults(baseDefaults, that.typeName, otherGrades);
+            var newDefaults = fluid.copy(fluid.getGradedDefaults(baseDefaults, that.typeName, otherGrades));
             var defaultsBlock = fluid.findMergeBlocks(shadow.mergeOptions.mergeBlocks, "defaults")[0];
             defaultsBlock.source = newDefaults;
             shadow.mergeOptions.updateBlocks();
@@ -427,6 +445,7 @@ var fluid_1_5 = fluid_1_5 || {};
             optionsStrategy, shadow.invokerStrategy.strategy, shadow.memberStrategy.strategy, eventStrategy]};
             
         fluid.computeDynamicGrades(that, shadow, strategy, shadow.mergeOptions.mergeBlocks);
+        fluid.distributeOptions(that, strategy);
         
         return shadow.getConfig;
     };
@@ -758,7 +777,10 @@ var fluid_1_5 = fluid_1_5 || {};
         var list = [];
         fluid.each(mergeRecords, function (value, key) {
             value.recordType = key;
-            if (key !== "demands") {
+            if (key === "distributions") {
+                list.push.apply(list, value);
+            }
+            else if (key !== "demands") {
                 if (!value.options) { return; }
                 value.priority = fluid.mergeRecordTypes[key];
                 if (value.priority === undefined) {
@@ -784,7 +806,7 @@ var fluid_1_5 = fluid_1_5 || {};
 
     // unsupported, NON-API function - used from Fluid.js
     fluid.generateExpandBlock = function (record, that, mergePolicy) {
-        var expanded = fluid.expandOptions(record.options, that, mergePolicy, null, {defer: true});
+        var expanded = fluid.expandOptions(record.options, record.contextThat || that, mergePolicy, null, {defer: true});
         expanded.priority = record.priority;
         expanded.recordType = record.recordType;
         return expanded;
@@ -892,7 +914,10 @@ var fluid_1_5 = fluid_1_5 || {};
         }
         var demands = fluid.makeArray(demandspec.args);
         
-        var upDefaults = fluid.defaults(demandspec.funcName); // I can SEE into TIME!!
+        var upDefaults = fluid.defaults(demandspec.funcName);
+        
+        var distributions = upDefaults && parentThat ? fluid.receiveDistributions(parentThat, upDefaults.gradeNames, options.memberName) : [];
+        
         var argMap = upDefaults? upDefaults.argumentMap : null;
         var inferMap = false;
         if (!argMap && (upDefaults || (options && options.componentRecord)) && !options.passArgs) {
@@ -930,9 +955,21 @@ var fluid_1_5 = fluid_1_5 || {};
             if (demandspec[name] !== undefined && localRecord[name] === undefined) {
                 localRecord[name] = demandspec[name];
             }
+            if (name !== "options") {
+                for (var i = 0; i < distributions.length; ++ i) { // Apply non-options material from distributions (FLUID-5013)
+                    if (distributions[i][name] !== undefined) {
+                        localRecord[name] = distributions[i][name];
+                    }
+                }
+            }
         });
+        for (var i = 0; i < distributions.length; ++ i) {
+            if (distributions[i].type !== undefined) {
+                demandspec.funcName = distributions[i].type;
+            }
+        }
         
-        var mergeRecords = {};
+        var mergeRecords = {distributions: distributions};
         
         if (options.componentRecord !== undefined) {
             // Deliberately put too many things here so they can be checked in expandComponentOptions (FLUID-4285)
@@ -1002,6 +1039,16 @@ var fluid_1_5 = fluid_1_5 || {};
         };
     };
     
+    /** Instantiate the subcomponent with the supplied name of the supplied top-level component. Although this method
+     * is published as part of the Fluid API, it should not be called by general users and may not remain stable. It is
+     * currently the only mechanism provided for instantiating components whose definitions are dynamic, and will be 
+     * replaced in time by dedicated declarative framework described by FLUID-5022.
+     * @param that {Component} the parent component for which the subcomponent is to be instantiated
+     * @param name {String} the name of the component - the index of the options block which configures it as part of the
+     * <code>components</code> section of its parent's options
+     */
+     // NB "directArgs" is now disused by the framework
+    
     fluid.initDependent = function (that, name, directArgs) {
         if (that[name]) { return; } // TODO: move this into strategy
         directArgs = directArgs || [];
@@ -1045,6 +1092,7 @@ var fluid_1_5 = fluid_1_5 || {};
         that[name] = instance;
         fluid.fireEvent(instance, "events.onAttach", [instance, name, that]);
         fluid.popActivity();
+        return instance;
     };
     
     // unsupported, non-API function
@@ -1199,12 +1247,6 @@ outer:  for (var i = 0; i < exist.length; ++i) {
         return matches;   
     };
 
-    fluid.getMembers = function (holder, name) {
-        return fluid.transform(holder, function(member) {
-            return fluid.get(member, name);
-        });
-    };
-
     // unsupported, non-API function
     fluid.locateDemands = function (parentThat, demandingNames) {
         var matches = fluid.locateAllDemands(parentThat, demandingNames);
@@ -1261,6 +1303,9 @@ outer:  for (var i = 0; i < exist.length; ++i) {
                 // Resolve this material late, to deal with cases where the target has only just been brought into existence
                 // (e.g. a jQuery target for rendered material) - TODO: Possibly implement cached versions of these as we might do for invokers
                 var resolvedThis = fluid.expandOptions(recthis, that);
+                if (typeof(resolvedThis) === "string") {
+                    resolvedThis = fluid.getGlobalValue(resolvedThis);
+                }
                 if (!resolvedThis) {
                     fluid.fail("Could not resolve reference " + recthis + " to a value"); 
                 }
@@ -1654,7 +1699,7 @@ outer:  for (var i = 0; i < exist.length; ++i) {
     fluid.fetchExpandChildren = function (target, source, mergePolicy, miniWorld, options) {
         if (source.expander /* && source.expander.type */) { // possible expander at top level
             var expanded = fluid.expandExpander(target, source, options);
-            if (fluid.isPrimitive(expanded) || (fluid.isArrayable(expanded) ^ fluid.isArrayable(target))) {
+            if (fluid.isPrimitive(expanded) || fluid.isDOMish(expanded) || (fluid.isArrayable(expanded) ^ fluid.isArrayable(target))) {
                 return expanded;
             }
             else { // make an attempt to preserve the root reference if possible
